\name{deprecated}

\alias{deprecated}
\alias{defunct}
\alias{basePath}
\alias{srapply}
\alias{readAligned,BamFile-method}

\title{Deprecated and defunct functions}

\description{

  These functions were introduced but are now deprecated or defunct.

}

\details{

  Defunct functions: \itemize{

    \item{\code{srapply}}{. Use the BiocParallel package instead.}

    \item{\code{readAligned,BamFile-method}}{. Use the GenomicAlignments
      package instead.}

    \item{\code{basePath()}}{}

  }

}

\keyword{manip}
