\name{deprecated}

\alias{deprecated}
\alias{defunct}
\alias{basePath}

\title{Deprecated and defunct functions}

\description{

  These functions were introduced but are now deprecated or defunct.

}
\usage{
basePath(object, ...)
}

\arguments{

  \item{object}{For \code{basePath}, and object of class
	\code{ExperimentPath}.}

  \item{...}{Additional arguments.}
}

\keyword{manip}
