\name{ShortReadBase-package}
\alias{ShortReadBase-package}
\alias{.ShortReadBase-class}
\alias{show,.ShortReadBase-method}
\alias{append,.ShortReadBase,.ShortReadBase-method}
\alias{.Solexa-class}
\alias{.Roche-class}
\docType{package}
\title{

  FASTQ input and manipulation.

}
\description{

  This package implements sampling, iteration, and input of FASTQ
  files. The package includes functions for filtering and trimming
  reads, and for generating a quality assessment report. Data are
  represented as DNAStringSet-derived objects, and easily manipulated
  for a diversity of purposes.  The package also contains legacy support
  for early single-end, ungapped alignment formats.

}
\details{

  See \code{packageDescription('ShortRead')}

}
\author{
  Maintainer: Martin Morgan <mtmorgan@fhcrc.org>
}
\keyword{package}
