\name{ShortReadBase-package}
\alias{ShortReadBase-package}
\alias{.ShortReadBase-class}
\alias{show,.ShortReadBase-method}
\alias{append,.ShortReadBase,.ShortReadBase,ANY-method}
\alias{.Solexa-class}
\alias{.Roche-class}
\docType{package}
\title{

  Base classes and methods for high-throughput short-read sequencing data.

}
\description{

  Base classes, functions, and methods for representation of
  high-throughput, short-read sequencing data.

}
\details{

  See \code{packageDescription('ShortRead')}

}
\author{
  Maintainer: Martin Morgan <mtmorgan@fhcrc.org>
}
\keyword{package}
